Solve both parts of Day 1
This is a working version of both parts of Day 1 (in Rust). #NathanielBumppo was a _huge_ help at the end, suggesting the use of `select_nth_unstable()`, which did a nifty job of helping with both parts.

We'd had a nice version of the first part before that (lots of help from #esitsu and #ikopor). Kind of made a mess of the part one testing, and I'll need to clean that up, but we were running late and there you are.