Refactored Day 01
This is a refactoring of the working version of Day 01 that we had. I moved the key data processing into its own function (`process_elves`) so we could run tests on it. This returns a `Result<[usize; 3]>, which I think is a good use of Rust arrays since we know there should be exactly three values.