Got part one of Day 07 working
This worked the first time we ran it!

We don't seem to ever use the string associated with a `dir` entry, so maybe we could remove that.

If the `directory_name` in `InputLine::Cd` were an enum with three variants, then the type checker would be able to help up out some more in, e.g., `handle_cd()`.