Get Day 07, part 2 working
I had expected this to be a big change from part 1, but it was in fact a pretty tiny change, and we got the whole thing done in under 15 minutes! Really glad we didn't end up overcomplicating the first part, as what we did do put us exactly where we needed to be for this part.

Thanks to ikopor@Twitch for pointing out the use of `copied()` as a way of turning `Option<&T>` into `Option<T>`. That's still the kind of stuff that I never think about.