*Ugly* solution to part 1 of Day 8
This is really pretty atrocious, TBH. It works, but there is a huge amount of duplication and it just makes me want to cry a lot.

Wgaffa@Twitch suggested just rotating the Forest for each of the four directions (which is how they solved it) as a way of avoiding the duplication in the four directions. I actually like that idea, and might consider rewriting this to use that approach.