A complete mess attempting to rewrite AoC Day 08, part 1
We spent an entire 2 hour streaming session on this, and it's a disaster. I think we'll pitch all this and then try to seriously attempt either Wgaffa@Twitch's idea of rotating the matrix repeatedly, or esitsu@Twitch's idea of having an iterator that goes over the trees above, below, to the left, and to the right of a given tree.