Rewrite Part 1 of Day 08 more cleanly
This is a major rewrite of part 1 of Day 08 which dramatically simplifies the logic and improves the readability.

`Tree`s are now just `u8` and don't store `visible`, which means that `Forest`s don't have to be mutable, which great simplifies the sharing, ownership, and lifetime issues.

The `Direction` enumerator is slightly simpler, and we use `EnumIter` to be able to loop over the four directions.

This uses esitsu@Twitch's idea of treating `Tree` independently, checking all the trees in the four directions for a taller tree.

We had some issues in the `is_taller_than()` method getting the `impl Iterator` argument type sorted out, but once that was done things were pretty clean.

Wgaffa@Twitch suggested the nice use of `zip()` and `repeat()` to clean up the creation of the necessary (x, y) pairs.

The built-in functions `any()`, `all()`, `filter_map()`, and `flat_map()` were all useful in avoiding explicit looping.