Set up part 2, Day 08, and add `ForestIter` enum
This copies over the part 1 code.

While I was here, I also decided to explore esitsu@Twitch's idea of having an `enum` for the iterators over trees in the forest as a way of handling the fact that we have different iterator types.

So I introduced the `ForestIter` type, moved `neighbors()` to be a method on the `Direction` enum type, and then added an `all()` method to `ForestIter` that just calls `iter.all(f)` for both of the iterator types.