Make `ForestIterator` an actual _iterator`
This builds on a suggestion from esitsu@Twitch that instead of just implementing the `all` method, I should have `ForestIterator` actually implement the `Iterator` trait (and thus the `next()` method), and then I'd get `all` and a zillion other things for free.

I also tried to generalize `ForestIterator` some while I was there. Not 100% sure that I think that's a huge improvement, but I think it's at least a bit of an improvement?