Get Day 09, part 2 working
This passes the tests for part 2. Spent a while flailing because I didn't realize that some knots can now move diagonally, and didn't handle that case. We added some tests that indicated what the problem was, and then the fix was reasonably straightforward.

While there, however, the pattern matching in `update_knot()` got fairly ugly and repetative. So we used some conditional match clauses to clean that up quite a bit. esitsu@Twitch also pointed out the nice `Position {y, ..}` syntax for only matching on the parts you care about.