Start Day 11, part 1
This makes a good start on part 1 of Day 11, but there's still two key (and non-trivial) functions still to be filled in. We got all the parsing done, and after several bouts of flailing got what I think is a reasonable structure for `MonkeyState` and the looping, including a nifty use of `cycle`.

There are _piles_ of `unwrap()` in all the parsing code. The parsing is pretty complex, and I ended up just cheating and not really dealing with error cases all that cleanly. I _might_ come back to clean that up, but no promises. :-)