Switch from `clone()` to `mem::take()`
This changes the setup of `MonkeyState::items` so that we take ownership of the item vectors in the initial monkeys using `std::mem::take()` instead of cloning them. After this, every monkey has no items, and their initially assigned items are in `MonkeyState::items`. Very nifty, and avoids making new, unnecessary, `Vec`s.