Solve part 1 of Day 11
We dropped my use of a separate field to keep track of items, and just used the monkeys for that.

I thought I might need a `VecDeque`, but in the end didn't.

Clippy didn't like my use of `match` on a boolean in `process_item()`, but I liked it so I told Clippy to allow it.

The ownership and borrowing in `process_item()` was interesting. We had an immutable borrow at the top of the function, and needed a mutable reference at the bottom, so we needed to make sure we had finished with the `&Monkey` so the immutable borrow could be released before doing the mutable steps.