Solve part 2 of Day 11!
The change to the logic was really simple (remove the `/ 3` code), but we had to figure out how to deal with the huge numbers that resulted.

I was all set to bring in a bigint crate, but ikopor@Twitch and esitsu@Twitch encouraged thinking of other ways to limit the size of the numbers. Some very rusty number theory suggested we could mod by the product of all the `Monkey::test_value`s without affecting the logic, and it worked!

I tried just using `u128` without that trick, and that didn't work. With the trick, we were good with `u64`, but `u32` wasn't big enough.