Tidy up Part 1, Day 12
This change `x`/`y` to `row`/`col`, which I think makes things more readable and less likely to be buggy. 

I removed all references to `end` since (at least for part 1), we don't need it. It'll be a minor bummer if we do need it for part 2, but not that hard to put back.

I extracted a new `accessible_locations` function, which hopefully makes things a little more readable. Alternatively, I could put all that back in `shortest_path_length` and replace the `for` loop with a call to `for_each`. Not sure how I feel about that.