Implement logic for Part 1 of Day 13
This successfully solves Part 1 of Day 13. There was a lot of discussion of my `Packet::partial_cmp` implementation, which folks think is more complex than needed and involves the creation (& dropping) of a bunch of small vectors. esitsu@Twitch had a different idea that they posted something about on Discord (which I haven't yet looked at), so we might want to use that to tidy this up.

Also, NathanielBumppo@Twitch's solution did a ice job of using the parse inside `TryFrom` implementations in a way that's cleaner that what I have, so I might want to follow that model.