Solve Part 1 of Day 15
I did some math on the (which really ought to be in a comment on `SensorBeacon::row_range`) where we can, for each sensor, compute a range that represents the intersection of that sensors coverage area and the target row.

The `range_union_find` crate was nice for creating a union of those ranges. It also let us compute the total coverage of the range union, which is nearly the right answer.

The last bit was to find all the beacons on that row, and subtract them from the count since we're looking for places where there can not be a beacon (and those represent places there definitely is a beacon).

It turns out that its possible (and apparently common) for multiple sensors to detect and report the same beacon, so we used the `unique()` function from `itertools` to get a unique set of beacon points before filtering and counting.