Implement the parsing for Day 16
This uses `nom` to implement the parsing for Day 16. There's a lot I like about `nom`, but importing the right thing continues to be a challenge, and the whole `IResult` thing is still kinda confusing when converting from that to `anyhow::Result`. We ended up with the `extract_valve`/`valve` combo, which seemed a bit awkward. I wonder if I want a generic helper function that converts `IResult`s into `anyhow::Result`?

With this we're able to parse out all the `Valve` structs. We now need to construct the graph datatype that this represents. Actually solving the problem will involve some complex algorithms fun (maybe A*?), so I'll need to do a little homework before the next session so I'm ready to implement that.