Add a `BitSet` type, with a few tests
MizardX@Twitch had suggested using a `u64` bit set instead of a `HashSet` like I was using since the bit set would be a lot easier to clone (and we'll need to clone it a lot).

I explored using https://github.com/contain-rs/bit-set, but it seems wired to only use `u32` for its bits, and we need `u64`. :-(

So I ended up making my own. I really didn't trust myself on that, so I actually added some tests to make sure that it's doing the right thing.