Set things up for the dynamic programming solution for Day 16
Pretty much the whole session today was spent building the bridge between the basic parsing we did last week, and the actual dynamic programming algorithm that I hope will provide a solution.

There was a lot of flailing about the difference between a valve's name (e.g., "AA") and its number (e.g., 7). We need the number for things like the bit sets, and we need it to be a "small" number in the range 0..64 so that we can use a `u64` to hold the bit set.

In the end I created a new `NumberedValve` type that wraps `Valve` and adds the number. We then set those up after we've parsed all the valves and put the in the `Cave` struct.

We've started `Cave::max_release`, which is where the key logic for the dynamic programming solution will go. I have a question about a particular optimization that I haven't resolved; not sure whether we need that optimization or not. I think we'll do the example case first and see how many cases that would prune out there before we worry about it in the full test case.