Get Part 1 of Day 16 working!
I created a `State` type to hold the state of the search.

There was a lot of flailing in getting `max_release()` finished, much of which was related to ownership and (eventually) cloning. If I cared a lot about the performance (which I don't at the moment), we would want to look into avoiding (or speeding up) a lot of that cloning. I suspect that if we converted all valve names from strings to numbers that would improve the situation quite a lot.