Start Part 2 of Day 16
I think this gets us _close_, but not quite there. What I've done is implement _one_ total for each minute instead of _two_ moves per minute, so we're low.

Just for fun, I tried bumping it up to 54 minutes and dividing the remaining time by 2 when we open a valve, but that overshoots.That basically allows one of the two "players" to have more than 26 moves if that's advantageous, and that's not cool.

I think we need to create the cross product of the possible states, which will essentially square the out degree from each state, and multiply the number of possible states by the number of values (~60). I'm _hoping_ we can just do that and have it work, but we'll see.