Finish Part 2, Day 16
This wraps up Part 2 of Day 16. The code isn't very pretty, though; there's duplication and tons of cloning and general hackery. But it works. You do have to be patient, though, as it takes nearly 18 minutes to complete on the full input.

There are parts of it which are nicer than before, TBH. I really like `moves()` and `apply_moves()`, for example.

But there's so much icky in other places, that I kinda want to cry.