Fix some off-by-one errors in vertical spacing
I had two important off-by-one errors in the vertical spacing. The first is that we wanted the first block to "land" at level _one_, not zero, so we changed the check in `drop()` to be `y > 1` instead of `y > 0`.

The other problem was that I wasn't placing new rocks "high enough" in the cavern, and I needed to add 4 (instead of 3) to the `highest_rock_point` when a new rock was added to the world.