Implement basic framework for Part 1, Day 19
This gets the basic structure for Day 19 (Part 1) written, but the memoization totally failed to do what I wanted (& we needed) and things will run forever in the current setup.

I'm obviously going to have to work out something fancier for the dynamic programming/memoization to make this run in a reasonable amount of time.