Get Part 1 of Day 19 working
This took a lot of work to get it to finish in a timely manner, but we eventually got there with lots of help from people on Twitch.

The big change is adding conditions to the `make_x_robot` functions to return `None` if there seem to be enough of that kind of robot already in the world.