Refactor Day 20, Part 1
The code for Day 20, Part 1, is pretty ugly, and this hopefully makes it at least somewhat better. I created a new `MovedElement` type that wraps `Element` and adds new `current_position` and `new_position` fields. My hope is that I can now do some proper testing on those, perhaps using the `quickcheck` library.

Most of the logic is now is the `MovedElement::new()` method, which also manages to encapsulate several of the local variables and weird type conversion stuff there. It's still a little icky, but it's less icky than it was.

The other substantial bit of logic is in `move_element`. After calling `MovedElement::new()`, it sorts out the rotations based on the results in the returned `MovedElement`. Looking at that logic, that really should be in a method in `MovedElement` since it's _full_ of calls that reach into the `MovedElement` to get various elements.