Get Part 2 of Day 20 working
The three big changes we had to make were:

- Changing all the value types from `i16` to `i64`
- Multiplying all the values by the big constant specified in the problem
- Mixing 10 times

The first was a simple search and replace. The second was just adding a multiplication, although that introduced a type problem which we resolved by specifying the particular type (`i64`) to parse the input numbers to. The third was just a simple loop. And then it worked!