Start the solution for Day 21
This actually implements the key logic for Day 21, and now I think all we need to do is the parsing.

The current implementation of `Monkeys::get_value` has a semi-nasty `clone()` call that is likely to cost us something, but I think it just adds a constant factor to the whole process, and there are only a little over 2K variables to look up, so my suspicion is that this will work as is.