Change the keys from `String` to `Rc<str>`
This follows a suggestion in [this YouTube video]( https://www.youtube.com/watch?v=A4cKi7PTJSs), where we use `Rc<str>` instead of `String` for our keys. This makes the keys immutable (which they're supposed to be as per `HashMap`), and makes cloning them really cheap, as opposed to cloning `String`s.