Solve Part 2 of Day 21
We ended up needing to do some (simple) symbolic algebra to solve this one in a timely fashion (thanks to my viewers for talking me into that!

We needed rational numbers, so I pulled in the `num-rational` crate for that.

Once a decision had been made, it wasn't _too_ bad to get it all to work. MizardX@Twitch was a huge help with the design (since they'd already solved it), and Vincent_SIFTD@Twitch also caught a lot of issues and made numerous excellent suggestions.