Rename some types; derive some traits
This adds `Eq` and `PartialEq` to some types so we can do equality tests on them, and `Copy` and `Clone` to others so they can just be copied and I don't have to worry about a bunch of references.

I renamed `Move` to `Action`, and `Directions` to `Actions`. The latter of those avoided confusion with the new `Direction` type as a cardinal direction.