Pass references to `Position`s (sometimes)
This changes things so that we pass references to `Position`s in at least some places where we were (I think) forcing copies to be made.

I'm not entirely sure this is an improvement, or that it's the best way to do things if it is an improvement. I think I'd need to map out more carefully how all the `Position` values are being created, passed, and returned.