Got Part 2 of Day 22 finished!
Huge thanks to MizardX@Twitch for the basic logic of moving from one cube face to the next, and for being very in explaining it to me over and over.

We implemented `Add` and `Sub` on `FacePosition` to making it a little cleaner to add movement vectors to positions.

The `Position`, `FacePosition`, and `You` structs have a lot of overlap and almost certainly could be refactored in useful ways.

There was an error where I was still using an independent `direction` argument in places instead of using the `direction` field in `Position` and `FacePosition`. Once I cleaned that up, it worked!

There was one broken test that was easy to fix.