Implement the structure for `dijkstra`
This sets up the structure for the `pathfinding` implementation of `dijkstra`'s shortest path algorithm.

I addd a `Node` type that has a `Pos` (our current location` and `time` (the number of steps/minutes since the beginning). We'll use the `time` to go "backwards" in time, looking for blizzards in the initial map that will (over that time) move into "blocking" positions.

I also added `successors` and `finished` to `Map`; these provide the logic needed by `dijkstra`. Implementing `finished()` was trivial, and hopefully `successors()` won't be too hard when we return to this next week.