Change map to hold at most on blizzard at a location
This changes the `Map` so that it holds at most one `Blizzard` at any given location.

This kinda broke the `Display` implementation for `Map`; I've got it working, but I think what we really want is a version that takes a `time` and prints the `Map` as it would appear at that time in the future. If we did that, we'd need to bring back the possibility of multiple `Blizzards` at a given `Pos`. For now, though, this works.