Fix gap in `successors`
I forgot that we need to include "sitting still" as an option in the `successors` function, so we fixed that with a nice use of `.chain(once(pos))`.

We also changed `successors` to return an iterator instead of returning constructing a `Vec`, which would then be immediately turned into an iterator by `astar`. That created a few lifetime and ownership issues that I flailed on for a bit, but we got there in the end. A key piece is that we made `Node` implement `Copy` (`Pos` already did), and then using `move` on several of the closures effectively copied values into the closures so we wouldn't have to worry about the ownership or lifetime issues.

We did have to to add `+ '_` at the end of the return type to make it clear that the lifetime of the return value should be the same as that as `&self`. Before we changed `Node` to be `Copy`, we needed specific lifetimes because we had to make it clear that `&self`, `Node`, and the return value all have the lifetime `a`.